+++ src/components/layout/PanelChange/index.tsx
@@ -16,7 +16,8 @@ import { useTranslation } from 'react-i18next';
 import en from 'javascript-time-ago/locale/en.json';
 import { getContract } from 'utils/getContract';
 import { rexReferralAddress } from 'constants/polygon_config';
-import { referralABI, streamExchangeABI } from 'constants/abis';
+import { referralABI } from 'constants/ABIs/referralABI';
+import { streamExchangeABI } from 'constants/ABIs/streamExchange';
 import { Coin } from 'constants/coins';
 import { FlowTypes } from 'constants/flowConfig';
 import { getShareScaler } from 'utils/getShareScaler';
